% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp3d.R
\name{mp3d}
\alias{mp3d}
\title{creates 3D HOV plot from raw microperimetry data file}
\usage{
mp3d(filepath)
}
\arguments{
\item{filepath}{file path to microperimetry file}
}
\description{
creates 3D HOV plot from raw microperimetry data file
}
